Add dial timeout by default
This sets dialing timeout on SSH calls. It could happen that if the
remote part doesn't shutdown the SSH connection properly, the connection
on golang could sit on ssh.Wait().

Setting a Dial timeout allows the client to poke with keep-alives and if
the remote part doesn't answer in a timeframe the connection is closed.

Signed-off-by: Ettore Di Giacinto <edigiacinto@suse.com>