Improve pod status check in NamespaceWithReadyPod
Checking that the pod's status is Ready is not enough, we have to be
sure that Started is set to as well as checking that containers are
really in Running state. Otherwise we validate that the Pods are OK even
if they are still in creating/starting process, which can take some
times to end.

Signed-off-by: Loic Devulder <ldevulder@suse.com>